+++ examples/common-js.js
@@ -1,0 +1,23 @@
+import { hash160 } from '@cashscript/utils';
+import {
+  deriveHdPrivateNodeFromSeed,
+  deriveHdPath,
+  secp256k1,
+  encodeCashAddress,
+} from '@bitauth/libauth';
+import bip39 from 'bip39';
+
+// This is duplicated from common.ts because it is not possible to import from a .ts file in p2pkh.js
+
+// Generate entropy from BIP39 mnemonic phrase and initialise a root HD-wallet node
+const seed = await bip39.mnemonicToSeed('CashScript Examples');
+const rootNode = deriveHdPrivateNodeFromSeed(seed, true);
+const baseDerivationPath = "m/44'/145'/0'/0";
+
+// Derive Alice's private key, public key, public key hash and address
+const aliceNode = deriveHdPath(rootNode, `${baseDerivationPath}/0`);
+if (typeof aliceNode === 'string') throw new Error();
+export const alicePub = secp256k1.derivePublicKeyCompressed(aliceNode.privateKey);
+export const alicePriv = aliceNode.privateKey;
+export const alicePkh = hash160(alicePub);
+export const aliceAddress = encodeCashAddress('bchtest', 'p2pkh', alicePkh);