+++ packages/cashc/test/test-utils.ts
@@ -1,13 +1,14 @@
 import fs from 'fs';
-import path from 'path';
+import { URL } from 'url';
+import urlJoin from 'url-join';
 
-export function getSubdirectories(directory: string): string[] {
+export function getSubdirectories(directory: URL): string[] {
   return fs.readdirSync(directory)
-    .filter((fn) => fs.statSync(path.join(directory, fn)).isDirectory());
+    .filter((fn) => fs.statSync(new URL(urlJoin(directory.toString(), fn))).isDirectory());
 }
 
-export function readCashFiles(directory: string): { fn: string, contents: string }[] {
+export function readCashFiles(directory: URL): { fn: string, contents: string }[] {
   return fs.readdirSync(directory)
     .filter((fn) => fn.endsWith('.cash'))
-    .map((fn) => ({ fn, contents: fs.readFileSync(path.join(directory, fn), { encoding: 'utf-8' }) }));
+    .map((fn) => ({ fn, contents: fs.readFileSync(new URL(urlJoin(directory.toString(), fn)), { encoding: 'utf-8' }) }));
 }